#1047 P2 (step 2): extract session/entry.rs (public-facing data types)
Second step of the session structural split. Lifts the public-facing
session data types from session/mod.rs into session/entry.rs.

What moved into session/entry.rs (132 LOC):
- pub(crate) struct SessionDecision
- pub(crate) struct SessionMetadata
- pub(crate) struct SessionLookup
- pub(crate) struct ForwardSessionMatch
- pub(crate) enum SessionOrigin + impl SessionOrigin
- pub(crate) enum SessionDeltaKind
- pub(crate) struct SessionDelta
- pub(crate) struct ExpiredSession

What stays in session/mod.rs:
- struct SessionEntry — file-private internal storage type whose
  fields are accessed directly by SessionTable's impl. Moving it would
  require widening every field to pub(in crate::session); not worth
  the visibility churn for this pass.

session/mod.rs adds `mod entry; pub(crate) use entry::*;` so
external callers reaching `crate::session::SessionDecision`,
`crate::session::SessionDelta`, etc. continue to resolve through the
glob re-export. No public API change.

Pure relocation — bodies byte-for-byte identical, no visibility
changes (everything was already pub(crate)).

LOC distribution after:
- session/mod.rs:  1,113 (was 1,234)
- session/key.rs:    138 (unchanged)
- session/entry.rs:  132 (new)
- session/tests.rs: 1,759 (unchanged)

cargo test --release: 865 passed, 0 failed.
